backend/requirements.txt
fastapi
uvicorn
boto3
google-auth
requests
pydantic
pydantic-settings
python-multipart